refactor: 상품 목록 뷰 LayoutManger 객체 분리
