+++ app/src/main/java/woowacourse/shopping/ui/products/ProductsActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.activity.viewModels
-import androidx.recyclerview.widget.GridLayoutManager
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityProductsBinding
 import woowacourse.shopping.ui.cart.CartActivity
@@ -57,25 +56,11 @@ class ProductsActivity : DataBindingActivity<ActivityProductsBinding>(R.layout.a
 
     private fun initViewBinding() {
         binding.productItemsContainer.adapter = productsAdapter
-        binding.productItemsContainer.layoutManager = createLayoutManager()
+        binding.productItemsContainer.layoutManager = ProductsLayoutManager(this, productsAdapter)
         binding.productItemsContainer.itemAnimator = null
     }
 
-    private fun createLayoutManager(): GridLayoutManager =
-        GridLayoutManager(this, GRID_LAYOUT_SIZE).apply {
-            spanSizeLookup =
-                object : GridLayoutManager.SpanSizeLookup() {
-                    override fun getSpanSize(position: Int): Int {
-                        val viewType = productsAdapter.getItemViewType(position)
-                        return when (ProductsItemViewType.entries[viewType]) {
-                            ProductsItemViewType.PRODUCT -> PRODUCT_LAYOUT_SIZE
-                            ProductsItemViewType.LOAD_MORE -> LOAD_MORE_LAYOUT_SIZE
-                        }
-                    }
-                }
-        }
-
-    private fun initObservers() {
+    fun initObservers() {
         viewModel.products.observe(this) { products ->
             productsAdapter.submitItems(products)
             viewModel.loadHasMoreProducts()
@@ -83,11 +68,5 @@ class ProductsActivity : DataBindingActivity<ActivityProductsBinding>(R.layout.a
         viewModel.hasMoreProducts.observe(this) { hasMore ->
             productsAdapter.updateHasMoreItem(hasMore)
         }
-    }
-
-    companion object {
-        private const val GRID_LAYOUT_SIZE: Int = 2
-        private const val PRODUCT_LAYOUT_SIZE: Int = 1
-        private const val LOAD_MORE_LAYOUT_SIZE: Int = 2
     }
 }
